Handle all restore flow navigation on InitialWalkthroughPage
